Updated readmefiles and name fixes
